requirements.txt
torch==2.0.1
torchvision==0.15.2
flask==3.0.0
opencv-python==4.8.0.76
numpy==1.24.3
matplotlib==3.7.2
scikit-learn==1.3.0
Pillow==9.5.0